requirements.txt
flask==3.0.2
openai==1.1.1
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.0.post1
flask-migrate==4.0.5
flask-mail==0.9.1
itsdangerous==2.1.2 